+++ services/redis/mgmt/2021-06-01/redis/CHANGELOG.md
@@ -1,2 +1,10 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. CommonPropertiesRedisConfiguration.AofBackupEnabled
+1. CommonPropertiesRedisConfiguration.Authnotrequired